Keep HTML head/frontmatter assembly in one helper module
html.ts still contained head-element rendering and frontmatter merge logic while
html.test.ts duplicated both helpers inline. Extracting them to a dedicated
module shrinks the orchestrator file and makes the tests exercise the real
helper behavior instead of shadow copies.

Constraint: Keep the slice internal to rendering/orchestrator and avoid changing HTML output behavior
Rejected: Fold additional shell-generation helpers into the same PR | broadens the rendering hotspot slice beyond one smell
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Add future head/frontmatter formatting changes in html-head.ts so html.ts stays focused on document assembly
Tested: Targeted html/version tests, fmt, lint, typecheck, verify:quick
Not-tested: Downstream consumer repos because this slice is internal to veryfront-code
