+++ deno.json
@@ -1,6 +1,6 @@
 {
   "name": "veryfront",
-  "version": "0.1.170",
+  "version": "0.1.171",
   "license": "Apache-2.0",
   "nodeModulesDir": "auto",
   "exclude": [
